根据作者反馈新内容修改译文3.2.5
作者反馈3.2.5解释新内容如下：
Within DevOps a distinction is made between the release process and the deployment process. The release process is about the correctness of the new version of the service while the deployment process is about distributing the new version of the service. The release of a new version of a service is based on the test results of the involved requirements.